Removed upload_to_database.py as not needed any more and added its function to database.py
